auth somewhat working
